Avoid an assert w/ maybeGenerateMore being called during generation
Was hitting an assertion in combination with pipes on
the MacBook. Presumably a pipe was calling read() on
one block while the generator was still busy filling
buffers. The MacPro's may have been too fast to hit
that.
I think it is perfectly fine that read()'s and primary
reading can interleave. Or is it??