Use #if os(linux) to workaround SR-2907
This ties Linux to Swift 3.0.1 Preview 2 and macOS to
Swift 3.0.0. Depending on whether SR-2907 is fixed in
Swift 3.0.1, we may drop this weirdness.

Since we have to use Swift 3.0.1 Preview 2 for Swift
anyways, lets use the `#if os(linux)` to switch between
the language versions.

Note: In case this isn't obvious, `#if swift(>=3.0.1)`
      does NOT work (SR-2908).