Bug fix: Readable Pipe Crash (#4)
Fix piping of empty `Readable`'s

When a callback based `Readable` w/o an actual callback (manual push) got piped, it was issuing an EOF push. This would close the pipe.

Instead just ignore the read request and wait for the user to push something to the readable (content or EOF).