Working on the HTTP message flow
A lot of this is still not quite right for
persistent connections. I think the core
flaw is that the IncomingMessage directly
deals with the socket. That should be
dealt with in the HTTP connection, the
IncomingMessage should be a plain stream
that is fed by the connection. As it needs
to decode chunked and all that.