Bugfix: Once emitter
This should fix an old bug in which once listeners
triggered a warning log that a 'concurrent' change
happens.

That was in onReadable calls which also ended the
stream. In that case the endReadable would flush
all read listeners (while they are called).

We now filter out all the once-listeners before
calling them.

TBD: not quite sure whether the onReadable should
be cancelled. Probably not.