Swift 4.2: BIG CHANGE on `init` of sockaddr's
Swift 4.2 has issues w/ overiding the `init` of C imported
structs.

In the future you have to be very careful w/ using the
default empty `init` of sockaddr related structs! They
will be 0 initialized and not have the proper domain etc.

We also provide a static `make` functions, but well.